+++ oov_tokenizer.py
@@ -1,32 +1,25 @@
 from keras.preprocessing.text import Tokenizer, maketrans
+import string
 
-# we do not want to filter anything in code.
-def base_code_filter():
-    return ""
 
-def text_to_word_sequence(text, filters=base_code_filter(), lower=True, split=" "):
-    ''' Tokenizes code. All consecutive alphanumeric characters are grouped into one token.
-    Thereby trying to heuristically match identifiers.
-    All other symbols are seen as one token.
-    Whitespace is stripped, except the newline token.
-    '''
-    if lower:
-        text = text.lower() #type: str
-    seq = []
-    curr = ""
-    for c in text:
-        if c.isalnum():
-            curr += c
-        else:
-            if curr != "":
-                seq.append(curr)
-                curr = ""
-            if not c.isspace() or c == '\n':
-                seq.append(c)
-    return [_f for _f in seq if _f]
+# a tokenizer that supports out of vocabulary tokens
+class OOVTokenizer(Tokenizer):
+    def base_filter(self):
+        f = string.punctuation
+        f = f.replace("'", '')
+        f += '\t\n'
+        return f
 
-# this class overrides methods to be able to use the text_to_word_sequence function above.
-class CodeTokenizer(Tokenizer):
+    # this is added to the class to support overriding in subclasses, like the code tokenizer
+    def text_to_word_sequence(self, text, filters=base_filter(), lower=True, split=" "):
+        '''prune: sequence of characters to filter out
+        '''
+        if lower:
+            text = text.lower()
+        text = text.translate(maketrans(filters, split*len(filters)))
+        seq = text.split(split)
+        return [_f for _f in seq if _f]
+
     def fit_on_texts(self, texts):
         '''Required before using texts_to_sequences or texts_to_matrix
 
@@ -37,7 +30,7 @@ class CodeTokenizer(Tokenizer):
         self.document_count = 0
         for text in texts:
             self.document_count += 1
-            seq = text if self.char_level else text_to_word_sequence(text, self.filters, self.lower, self.split)
+            seq = text if self.char_level else self.text_to_word_sequence(text, self.filters, self.lower, self.split)
             for w in seq:
                 if w in self.word_counts:
                     self.word_counts[w] += 1
@@ -55,6 +48,8 @@ class CodeTokenizer(Tokenizer):
         # note that index 0 is reserved, never assigned to an existing word
         self.word_index = dict(list(zip(sorted_voc, list(range(1, len(sorted_voc) + 1)))))
 
+        self.oov_token = len(sorted_voc) + 1
+
         self.index_docs = {}
         for w, c in list(self.word_docs.items()):
             self.index_docs[self.word_index[w]] = c
@@ -62,7 +57,7 @@ class CodeTokenizer(Tokenizer):
     def texts_to_sequences_generator(self, texts):
         '''Transforms each text in texts in a sequence of integers.
         Only top "nb_words" most frequent words will be taken into account.
-        Only words known by the tokenizer will be taken into account.
+        Words that are not known by the tokenizer will be replaced by self.oov_token token.
 
         Yields individual sequences.
 
@@ -71,13 +66,15 @@ class CodeTokenizer(Tokenizer):
         '''
         nb_words = self.nb_words
         for text in texts:
-            seq = text if self.char_level else text_to_word_sequence(text, self.filters, self.lower, self.split)
+            seq = text if self.char_level else self.text_to_word_sequence(text, self.filters, self.lower, self.split)
             vect = []
             for w in seq:
                 i = self.word_index.get(w)
                 if i is not None:
                     if nb_words and i >= nb_words:
-                        continue
+                        vect.append(self.oov_token)
                     else:
                         vect.append(i)
+                else:
+                    vect.append(self.oov_token)
             yield vect
